+++ ehrapy/_settings.py
@@ -54,7 +54,6 @@ class EhrapyConfig:  # pragma: no cover
         cache_compression: str | None = "lzf",
         max_memory=15,
         n_jobs: int = -1,
-        logfile: str | Path | None = None,
         categories_to_ignore: Iterable[str] = ("N/A", "dontknow", "no_gate", "?"),
         _frameon: bool = True,
         _vector_friendly: bool = False,
