+++ tests/utils/test_utils_available.py
@@ -1,4 +1,4 @@
-from ehrapy._utils_available import _check_module_importable, _shell_command_accessible
+from ehrapy._compat import _check_module_importable, _shell_command_accessible
 
 
 def test_check_module_importable_true():
